+++ base/unary-reduce-strided1d-dispatch-factory/lib/main.js
@@ -35,11 +35,13 @@ var UnaryStrided1dDispatch = require( './../../../base/unary-reduce-strided1d-di
 * @param {ArrayLikeObject<Function>} [table.fcns] - list of strided reduction functions which are specific to specialized input ndarray argument signatures
 * @param {ArrayLikeObject<StringArray>} idtypes - list containing lists of supported input data types for each ndarray argument
 * @param {StringArray} odtypes - list of supported output data types
-* @param {string} policy - output data type policy
+* @param {Object} policies - policies
+* @param {string} policies.output - output data type policy
+* @param {string} policies.casting - input ndarray casting policy
 * @throws {TypeError} first argument must be an object having valid properties
 * @throws {TypeError} second argument must be an array containing arrays of supported data types
 * @throws {TypeError} third argument must be an array of supported data types
-* @throws {TypeError} fourth argument must be a supported output data type policy
+* @throws {TypeError} fourth argument must be an object having supported policies
 * @throws {Error} first argument must be an object having valid properties
 * @returns {Function} function for performing a reduction on an ndarray
 *
@@ -50,12 +52,15 @@ var UnaryStrided1dDispatch = require( './../../../base/unary-reduce-strided1d-di
 *
 * var idt = dtypes( 'real_and_generic' );
 * var odt = idt;
-* var policy = 'same';
+* var policies = {
+*     'output': 'same',
+*     'casting': 'none'
+* };
 *
 * var table = {
 *     'default': base
 * };
-* var max = factory( table, [ idt ], odt, policy );
+* var max = factory( table, [ idt ], odt, policies );
 *
 * var xbuf = [ -1.0, 2.0, -3.0 ];
 * var x = new ndarray( 'generic', xbuf, [ xbuf.length ], [ 1 ], 0, 'row-major' );
@@ -73,12 +78,15 @@ var UnaryStrided1dDispatch = require( './../../../base/unary-reduce-strided1d-di
 *
 * var idt = dtypes( 'real_and_generic' );
 * var odt = idt;
-* var policy = 'same';
+* var policies = {
+*     'output': 'same',
+*     'casting': 'none'
+* };
 *
 * var table = {
 *     'default': base
 * };
-* var max = factory( table, [ idt ], odt, policy );
+* var max = factory( table, [ idt ], odt, policies );
 *
 * var xbuf = [ -1.0, 2.0, -3.0 ];
 * var x = new ndarray( 'generic', xbuf, [ xbuf.length ], [ 1 ], 0, 'row-major' );
@@ -95,8 +103,8 @@ var UnaryStrided1dDispatch = require( './../../../base/unary-reduce-strided1d-di
 * var bool = ( out === y );
 * // returns true
 */
-function factory( table, idtypes, odtypes, policy ) {
-	var f = new UnaryStrided1dDispatch( table, idtypes, odtypes, policy );
+function factory( table, idtypes, odtypes, policies ) {
+	var f = new UnaryStrided1dDispatch( table, idtypes, odtypes, policies );
 	setReadOnly( main, 'assign', assign );
 	return main;